requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alpaca-trade-api==3.2.0
attrs==25.3.0
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
cssselect2==0.8.0
cycler==0.12.1
deprecation==2.1.0
fonttools==4.57.0
frozenlist==1.6.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.1
msgpack==1.0.3
multidict==6.4.3
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pdfkit==1.0.0
pillow==11.2.1
propcache==0.3.1
pycparser==2.22
pydyf==0.11.0
pyparsing==3.2.3
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.1
requests==2.32.3
six==1.17.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.2
urllib3==1.26.20
weasyprint==65.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
yarl==1.20.0
zopfli==0.2.3.post1
